Add .gitignore file
A .gitignore file has been created to ignore files that should not be tracked in the repository. This includes a range of file types and directories associated with different environments, IDEs, and languages, predominantly Python and Java related. This will help prevent unnecessary files from being included in the repository and will keep the repo clean and focused on the source code.